Add function biocReposList()
Currently not in NAMESPACE as this is preliminary.  Packages
can call Biobase:::biocReposList() to obtain a character vector of all
Bioconductor package repositories appropriate for that version of Biobase.

This allows packages that want to do automagic package installation, like affy,
to use install.packages and not depend at all on reposTools (which is going
away real soon now).



git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/Biobase@14256 bc3139a8-67e5-0310-9ffc-ced21a209358