Improve handling of unknown args in eSet's initialize method
We snarf unknown arguments passed in via '...' into elements of
assayData.  With this patch, we first attempt to assign these
elements to slots and only if no such slot is found does the
argument end up in assayData.

The reason to do this is that it allows subclasses to rely on default
initialize method behavior where extra args to initialize are treated
as slot initializers.


git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/Biobase@21351 bc3139a8-67e5-0310-9ffc-ced21a209358