Use 'unsafeSetSlot' within eSet replacement methods.
This relies on the _current implementation_ of S4 generics making a
copy before entering the replacement method -- incoming objects have
exactly one reference, and hence can be modified 'in place'. This
saves 2-4 copies of the object, but is terrible programming practice
(relying on implementation detail; direct slot assignment, ...) and
the problem is only partially 'real' (when assayData is an environment
or lockedEnvironment, the 'big' data isn't being copied anyway).


git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/Biobase@24726 bc3139a8-67e5-0310-9ffc-ced21a209358