PUBLIC: o Added anyMissing(). o Added support for missing values of rowMedians().
DETAILS:
o Now rowMedians() has an optimized C-level implementation. There is one implementation for real values and one for integers. This way an integer matrix does not have to be coerced to a double (which would add 200% memory usage). Moreover, at the C-level there is also an argument 'hasna'; if hasna=FALSE, missing values are not looked for resulting in a faster processing. At the R level, this argument is currently hidden, but may be added if found useful.



git-svn-id: https://hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/Biobase@26650 bc3139a8-67e5-0310-9ffc-ced21a209358